refactor(requestformatter): moved formatting functions and associated types to a util folder
TODO - add

re# 1325
