+++ src/ui/sso/components/checkboxCredential.tsx
@@ -2,15 +2,15 @@ import React from 'react'
 import { StyleSheet, View, Text } from 'react-native'
 import { Spacing, Typography, Colors } from 'src/styles'
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons'
-import {
-  CredentialTypeSummary,
-  CredentialVerificationSummary,
-  IdentitySummary,
-} from 'src/actions/sso/types'
 import I18n from 'src/locales/i18n'
 import strings from 'src/locales/strings'
 import { IconToggle } from 'react-native-material-ui'
 import { reject, isEmpty } from 'ramda'
+import {
+  CredentialTypeSummary,
+  CredentialVerificationSummary,
+} from '../../../utils/interactionRequests/types'
+import { IdentitySummary } from '../../../actions/sso/types'
 
 const styles = StyleSheet.create({
   container: {
